+++ sdk/recoveryservices/arm-recoveryservices/CHANGELOG.md
@@ -1,5 +1,32 @@
 # Release History
+    
+## 5.1.0 (2022-07-22)
+    
+**Features**
 
+  - Added Interface AzureMonitorAlertSettings
+  - Added Interface ClassicAlertSettings
+  - Added Interface MonitoringSettings
+  - Added Interface PatchTrackedResource
+  - Added Interface PatchVault
+  - Added Interface ResourceCertificateAndAadDetails
+  - Added Interface ResourceCertificateAndAcsDetails
+  - Added Interface TrackedResource
+  - Added Interface Vault
+  - Added Interface VaultExtendedInfoResource
+  - Added Interface VaultPropertiesRedundancySettings
+  - Added Type Alias AlertsState
+  - Added Type Alias BackupStorageVersion
+  - Added Type Alias CrossRegionRestore
+  - Added Type Alias StandardTierStorageRedundancy
+  - Interface VaultProperties has a new optional parameter backupStorageVersion
+  - Interface VaultProperties has a new optional parameter monitoringSettings
+  - Interface VaultProperties has a new optional parameter redundancySettings
+  - Added Enum KnownAlertsState
+  - Added Enum KnownBackupStorageVersion
+  - Added Enum KnownCrossRegionRestore
+  - Added Enum KnownStandardTierStorageRedundancy
+    
 ## 5.0.1 (2022-04-29)
 
 **Features**